Refactor namespaces and update project structure
This commit refactors the codebase to change the namespace from `jobportal` to `Jobpost.API`, affecting controllers, services, models, and DTOs. The `Jobpost` class has been renamed to `Jobposting` for clarity. Interfaces `IOnboardingService` and `IPostService` have been updated to use the new DTOs. The `Program.cs` file and solution file have been modified to reflect the new project structure, including the addition of the `JobPost.API` project. Model classes and DTOs have also been updated to ensure consistency across the application.
