+++ JobPost.API/Controllers/PostController.cs
@@ -1,9 +1,9 @@
-﻿using jobportal.Database;
-using jobportal.Dtos;
+﻿using Jobpost.API.Database;
+using Jobpost.API.Dtos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace jobportal.Controllers
+namespace Jobpost.API.Controllers
 {
     [Route("/api/[controller]")]
     public class PostController:ControllerBase
