+++ JobPost.API/Database/AppDbContext.cs
@@ -1,9 +1,9 @@
-﻿using jobportal.Model;
+﻿using Jobpost.API.Model;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
 
-namespace jobportal.Database
+namespace Jobpost.API.Database
 {
     public class AppDbContext: IdentityDbContext<User>
     {
@@ -15,7 +15,7 @@ namespace jobportal.Database
         }
         public DbSet<Employee> Employee { get; set; }
         public DbSet<Employer> Employers { get; set; }
-        public DbSet<Jobpost> Jobposts { get; set; }
+        public DbSet<Jobposting> Jobposts { get; set; }
         public DbSet<Company> Company { get; set; }
         public DbSet<Application> Application { get; set; }
         public DbSet<User> User { get; set; }
