+++ JobPost.API/Dtos/PostJobDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace jobportal.Dtos
+namespace Jobpost.API.Dtos
 {
     public class PostJobDTO
     {
