+++ JobPost.API/Model/Application.cs
@@ -1,17 +1,17 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace jobportal.Model
+namespace Jobpost.API.Model
 {
-    public class Company
+    public class Application
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
+        [Required]
+        public Employee? employee { get; set; }
+        [Required]
+        public Jobposting? Job { get; set; }
 
-        [Required]
-        public string? CompanyName { get; set; }
-        [Required]
-        public string? CompanyAddress { get; set; }
     }
 }
