+++ JobPost.API/Model/Company.cs
@@ -1,17 +1,17 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace jobportal.Model
+namespace Jobpost.API.Model
 {
-    public class Employer
+    public class Company
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        [Required]
-        public User? AppUser{ get; set; }
-        [Required]
-        public Company? Company { get; set; }
 
+        [Required]
+        public string? CompanyName { get; set; }
+        [Required]
+        public string? CompanyAddress { get; set; }
     }
 }
