+++ JobPost.API/Model/Employer.cs
@@ -1,15 +1,17 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace jobportal.Model
+namespace Jobpost.API.Model
 {
-    public class Jobpost
+    public class Employer
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        public string? JobTile { get; set; }
-        public string? JobDescription { get; set; }
-        public Employer? Employer { get; set; }
+        [Required]
+        public User? AppUser{ get; set; }
+        [Required]
+        public Company? Company { get; set; }
+
     }
 }
