+++ JobPost.API/Model/Jobposting.cs
@@ -1,17 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace jobportal.Model
+namespace Jobpost.API.Model
 {
-    public class Application
+    public class Jobposting
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        [Required]
-        public Employee? employee { get; set; }
-        [Required]
-        public Jobpost? Job { get; set; }
-
+        public string? JobTile { get; set; }
+        public string? JobDescription { get; set; }
+        public Employer? Employer { get; set; }
     }
 }
