+++ JobPost.AppHost/JobPost.AppHost.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <Sdk Name="Aspire.AppHost.Sdk" Version="9.0.0" />
 
@@ -17,8 +17,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\jobportal\jobportal.csproj" />
+    <ProjectReference Include="..\JobPost.API\JobPost.API.csproj" />
     <ProjectReference Include="..\JobPost.ServiceDefaults\JobPost.ServiceDefaults.csproj" />
   </ItemGroup>
+
 
 </Project>
