+++ JobPostingsApp.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35818.85
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "jobportal", "jobportal\jobportal.csproj", "{8DB39695-DA76-4104-BB61-33D457A25F9B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JobPost.AppHost", "JobPost.AppHost\JobPost.AppHost.csproj", "{86CEE165-C4ED-75AD-4B0E-B8E27AB8F423}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JobPost.ServiceDefaults", "JobPost.ServiceDefaults\JobPost.ServiceDefaults.csproj", "{AB77B354-837C-C6D5-3F8E-9576B2C775D5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JobPost.API", "JobPost.API\JobPost.API.csproj", "{B5413F1C-69C1-CBC4-B2E3-DABA96B31327}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8DB39695-DA76-4104-BB61-33D457A25F9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8DB39695-DA76-4104-BB61-33D457A25F9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8DB39695-DA76-4104-BB61-33D457A25F9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8DB39695-DA76-4104-BB61-33D457A25F9B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{86CEE165-C4ED-75AD-4B0E-B8E27AB8F423}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{86CEE165-C4ED-75AD-4B0E-B8E27AB8F423}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{86CEE165-C4ED-75AD-4B0E-B8E27AB8F423}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +23,10 @@ Global
 		{AB77B354-837C-C6D5-3F8E-9576B2C775D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AB77B354-837C-C6D5-3F8E-9576B2C775D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AB77B354-837C-C6D5-3F8E-9576B2C775D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5413F1C-69C1-CBC4-B2E3-DABA96B31327}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5413F1C-69C1-CBC4-B2E3-DABA96B31327}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5413F1C-69C1-CBC4-B2E3-DABA96B31327}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5413F1C-69C1-CBC4-B2E3-DABA96B31327}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
